Make it possible to run cleanup after each test and fix issue with stop and subprocess forks

+++ src/unix/run-process.ts
@@ -7,6 +7,8 @@ export type ExitInformation = { code: number | null; signal: NodeJS.Signals | nu
 export class TimeoutError extends Error {}
 export class ExitBeforeOutputMatchError extends Error {}
 export class StopBecauseOfOutputError extends Error {}
+export class StandardStreamsStillOpenError extends Error {}
+export class ProcessNotRunningError extends Error {}
 
 export class RunProcess {
   public cmd: ChildProcess
@@ -15,12 +17,15 @@ export class RunProcess {
   public stdout: ChildProcess['stdout'] = null
   public stderr: ChildProcess['stderr'] = null
   public running: boolean
+  public stopped = false
   public stopReason: Error | null = null
   public detached: boolean
 
   private startPromise: Promise<void>
   private stopPromise: Promise<ExitInformation>
+  // TODO: Expose all throws to the error listener
   private errorListeners: Array<(err: Error) => void> = []
+  private exitListeners: Array<(code: number | null, signal: NodeJS.Signals | null) => void> = []
 
   public constructor(command: string, args?: string[], options?: Parameters<typeof spawn>[2]) {
     // Jest does not give access to global process.env so make sure we use the copy we have in the test
@@ -65,12 +70,19 @@ export class RunProcess {
       ? new Promise<void>(resolve => this.cmd.stderr?.on('end', resolve))
       : Promise.resolve()
 
-    this.stopPromise = Promise.all([this.startPromise, exitPromise, stdoutPromise, stderrPromise]).then(
-      result => result[1]
-    )
-    this.stopPromise.catch(() => {
-      // User might never bind to this promise if they are just starting a process not caring about if it runs on not
+    this.stopPromise = Promise.all([this.startPromise, exitPromise, stdoutPromise, stderrPromise]).then(result => {
+      this.stopped = true
+      return result[1]
     })
+    this.stopPromise
+      .then(res => {
+        for (const listener of this.exitListeners) {
+          listener(res.code, res.signal)
+        }
+      })
+      .catch(() => {
+        // User might never bind to this promise if they are just starting a process not caring about if it runs on not
+      })
   }
 
   public async stop(sigKillTimeout = 3000, error?: Error): Promise<ExitInformation> {
@@ -83,7 +95,10 @@ export class RunProcess {
         }
         this.cmd.kill('SIGKILL')
       }
+    } else if (!this.stopped) {
+      throw new StandardStreamsStillOpenError('Process exited but standard steams are still open')
     }
+
     return await this.stopPromise
   }
 
@@ -116,36 +131,87 @@ export class RunProcess {
     return await this.stopPromise
   }
 
-  public waitForOutput(
+  public async waitForOutput(
     regex: RegExp,
     timeout = 0,
     outputs: Array<'stdout' | 'stderr'> = ['stdout', 'stderr']
   ): Promise<RegExpMatchArray> {
-    return new Promise((resolve, reject) => {
-      let timeoutHandle: NodeJS.Timeout | null = null
-      if (timeout) {
-        timeoutHandle = setTimeout(() => reject(new TimeoutError()), timeout)
-      }
+    const listeners: Array<['stdout' | 'stderr', (chunk: Buffer) => void]> = []
+    try {
+      return await new Promise((resolve, reject) => {
+        let timeoutHandle: NodeJS.Timeout | null = null
+        if (timeout) {
+          timeoutHandle = setTimeout(() => reject(new TimeoutError()), timeout)
+        }
 
-      let data = ''
-      for (const output of outputs) {
-        this[output]?.on('data', (chunk: Buffer) => {
-          data += chunk.toString('utf8')
-          const match = data.match(regex)
-          if (match) {
-            if (timeoutHandle) {
-              clearTimeout(timeoutHandle)
+        // Setup match listener for both stdout and stderr
+        let data = ''
+        for (const output of outputs) {
+          const matchListener = (chunk: Buffer): void => {
+            data += chunk.toString('utf8')
+            const match = data.match(regex)
+            if (match) {
+              if (timeoutHandle) {
+                clearTimeout(timeoutHandle)
+              }
+              resolve(match)
             }
-            resolve(match)
           }
+          listeners.push([output, matchListener])
+          this[output]?.on('data', matchListener)
+        }
+        // Throw if the process exist before finding the output
+        // eslint-disable-next-line @typescript-eslint/no-floating-promises
+        this.stopPromise.then(() => {
+          reject(this.stopReason ? this.stopReason : new ExitBeforeOutputMatchError(data))
         })
-      }
-      // Throw if the process exist before finding the output
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      this.stopPromise.then(() => {
-        reject(this.stopReason ? this.stopReason : new ExitBeforeOutputMatchError(data))
       })
-    })
+    } finally {
+      // Make sure we stop listening when match is found
+      for (const [output, listener] of listeners) {
+        this[output]?.removeListener('data', listener)
+      }
+    }
+  }
+
+  public removeListener(event: string | symbol, listener: (...args: unknown[]) => void): this {
+    switch (event) {
+      case 'exit': {
+        this.exitListeners = this.exitListeners.filter(l => listener != l)
+        break
+      }
+      case 'error': {
+        this.errorListeners = this.errorListeners.filter(l => listener != l)
+        this.cmd.removeListener(event, listener)
+        break
+      }
+      default: {
+        this.cmd.removeListener(event, listener)
+      }
+    }
+    return this
+  }
+
+  public removeAllListeners(event?: string | symbol | undefined): this {
+    switch (event) {
+      case 'exit': {
+        this.exitListeners = []
+        break
+      }
+      case 'error': {
+        this.errorListeners = []
+        this.cmd.removeAllListeners(event)
+        break
+      }
+      default: {
+        this.cmd.removeAllListeners(event)
+      }
+    }
+    return this
+  }
+
+  public listeners(event: string | symbol): Function[] {
+    return this.cmd.listeners(event)
   }
 
   public on(event: 'close', listener: (code: number, signal: NodeJS.Signals) => void): this
@@ -157,10 +223,7 @@ export class RunProcess {
   public on(event: string, listener: (...args: any[]) => void): this {
     switch (event) {
       case 'exit': {
-        // eslint-disable-next-line @typescript-eslint/no-floating-promises
-        this.stopPromise.then(res => {
-          listener(res.code, res.signal)
-        })
+        this.exitListeners.push(listener)
         break
       }
       case 'error': {
@@ -175,7 +238,10 @@ export class RunProcess {
     return this
   }
 
-  public kill(signal?: NodeJS.Signals): void {
-    this.cmd.kill(signal)
+  public kill(signal?: NodeJS.Signals): boolean {
+    if (!this.running) {
+      throw new ProcessNotRunningError()
+    }
+    return this.cmd.kill(signal)
   }
 }
